requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
httpx
jinja2
alembic
pytest
pytest-asyncio
#httpx-mock
python-dotenv
